requirements.txt
requests
psycopg2-binary
pandas
streamlit
plotly
python-dotenv
apache-airflow
sqlalchemy
schedule
python-dateutil